+++ fortran/src/H5Tf.c
@@ -34,7 +34,7 @@ nh5topen_c (hid_t_f *loc_id, _fcd name, int_f *namelen, hid_t_f *type_id)
 {
      int ret_value = -1;
      char *c_name;
-     int c_namelen;
+     size_t c_namelen;
      hid_t c_type_id;
      hid_t c_loc_id;
 
@@ -76,7 +76,7 @@ nh5tcommit_c (hid_t_f *loc_id, _fcd name, int_f *namelen, hid_t_f *type_id)
 {
      int ret_value = -1;
      char *c_name;
-     int c_namelen;
+     size_t c_namelen;
      hid_t c_type_id;
      hid_t c_loc_id;
      herr_t status;
@@ -993,7 +993,7 @@ nh5tget_member_name_c ( hid_t_f *type_id ,int_f* idx, _fcd member_name, int_f *n
   c_name = H5Tget_member_name(c_type_id, c_index);
   if (c_name == NULL ) return ret_value;
 
-  HD5packFstring(c_name, _fcdtocp(member_name), (int)strlen(c_name));
+  HD5packFstring(c_name, _fcdtocp(member_name), strlen(c_name));
   *namelen = (int_f)strlen(c_name);
   HDfree(c_name);
   ret_value = 0;
@@ -1017,7 +1017,7 @@ nh5tget_member_index_c (hid_t_f *type_id, _fcd name, int_f *namelen, int_f *idx)
 {
      int ret_value = -1;
      char *c_name;
-     int c_namelen;
+     size_t c_namelen;
      hid_t c_type_id;
      int c_index;
 
@@ -1254,7 +1254,7 @@ nh5tinsert_c(hid_t_f *type_id, _fcd name, int_f* namelen, size_t_f *offset, hid_
   hid_t c_type_id;
   hid_t c_field_id;
   char* c_name;
-  int c_namelen;
+  size_t c_namelen;
   size_t c_offset;
   herr_t error;
 
@@ -1396,7 +1396,7 @@ nh5tenum_insert_c(hid_t_f *type_id, _fcd name, int_f* namelen, int_f* value)
   int ret_value = -1;
   hid_t c_type_id;
   char* c_name;
-  int c_namelen;
+  size_t c_namelen;
   int_f c_value;
   herr_t error;
 
@@ -1443,7 +1443,7 @@ nh5tenum_nameof_c(hid_t_f *type_id, int_f* value, _fcd name, size_t_f* namelen)
   c_name = (char *)malloc(sizeof(char)*c_namelen);
   c_type_id = (hid_t)*type_id;
   error = H5Tenum_nameof(c_type_id, &c_value, c_name, c_namelen);
-  HD5packFstring(c_name, _fcdtocp(name), (int)strlen(c_name));
+  HD5packFstring(c_name, _fcdtocp(name), strlen(c_name));
   HDfree(c_name);
 
   if(error < 0) return ret_value;
@@ -1471,7 +1471,7 @@ nh5tenum_valueof_c(hid_t_f *type_id, _fcd name, int_f* namelen, int_f* value)
   int ret_value = -1;
   hid_t c_type_id;
   char* c_name;
-  int c_namelen;
+  size_t c_namelen;
   herr_t error;
   c_namelen = *namelen;
   c_name = (char *)HD5f2cstring(name, c_namelen);
@@ -1539,7 +1539,7 @@ nh5tset_tag_c(hid_t_f* type_id, _fcd tag, int_f* namelen)
   hid_t c_type_id;
   herr_t status;
   char* c_tag;
-  int c_namelen;
+  size_t c_namelen;
 
   c_namelen = *namelen;
   c_tag = (char *)HD5f2cstring(tag, c_namelen);
@@ -1576,7 +1576,7 @@ nh5tget_tag_c(hid_t_f* type_id, _fcd tag, int_f* taglen)
   c_tag = H5Tget_tag(c_type_id);
   if (c_tag == NULL ) return ret_value;
 
-  HD5packFstring(c_tag, _fcdtocp(tag), (int)strlen(c_tag));
+  HD5packFstring(c_tag, _fcdtocp(tag), strlen(c_tag));
   *taglen = (int_f)HDstrlen(c_tag);
   HDfree(c_tag);
   ret_value = 0;
